# cloud/specs/001-cloud-cabinet-control/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Cloud端储能柜集群管理系统 API
  version: 1.0.0
  description: Cloud端储能柜集群管理系统的RESTful API规范
  contact:
    name: Cloud System Team
servers:
  - url: https://api.example.com/api/v1
    description: 生产环境
  - url: http://localhost:8003/api/v1
    description: 开发环境
...
paths:
...
  /cabinets/{cabinet_id}/sync:
    post:
      tags:
        - Sensors
      summary: 接收Edge端数据同步
      parameters:
        - name: cabinet_id
          in: path
          required: true
          schema:
            type: string
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncDataRequest'
      responses:
        '200':
          description: 同步成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncDataResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /cabinets/{cabinet_id}/sensors/latest:
    get:
      tags:
        - Sensors
      summary: 查询储能柜最新传感器数据
      parameters:
        - name: cabinet_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LatestSensorDataResponse'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: Bearer {token} 或 API Key
  schemas:
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
    SuccessResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
        message:
          type: string
...
    SyncDataRequest:
      type: object
      required:
        - cabinet_id
        - timestamp
      properties:
        cabinet_id:
          type: string
        timestamp:
          type: string
          format: date-time
        sensor_data:
          type: array
          items:
            $ref: '#/components/schemas/SensorDataPoint'
        alerts:
          type: array
          items:
            $ref: '#/components/schemas/AlertData'
    SensorDataPoint:
      type: object
      required:
        - device_id
        - sensor_type
        - value
        - unit
        - timestamp
      properties:
        id:
          type: integer
        device_id:
          type: string
        sensor_type:
          type: string
        value:
          type: number
        unit:
          type: string
        timestamp:
          type: string
          format: date-time
        quality:
          type: integer
        synced:
          type: boolean
    AlertData:
      type: object
      required:
        - device_id
        - alert_type
        - severity
        - message
        - timestamp
      properties:
        id:
          type: integer
        device_id:
          type: string
        alert_type:
          type: string
        severity:
          type: string
        message:
          type: string
        value:
          type: number
        threshold:
          type: number
        timestamp:
          type: string
          format: date-time
        resolved:
          type: boolean
    SyncDataResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                synced_count:
                  type: object
                  properties:
                    sensor_data:
                      type: integer
                    alerts:
                      type: integer
                next_sync_recommended:
                  type: string
                  format: date-time
    LatestSensorDataResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                cabinet_id:
                  type: string
                timestamp:
                  type: string
                  format: date-time
                sensors:
                  type: array
                  items:
                    $ref: '#/components/schemas/SensorReading'
    SensorReading:
      type: object
      properties:
        sensor_type:
          type: string
        device_id:
          type: string
        value:
          type: number
        unit:
          type: string
        status:
          type: string
        quality:
          type: integer
        timestamp:
          type: string
          format: date-time
...
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: 未认证
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...